Using decimal in the JsonNumber struct instead of double to increase precision.

+++ Enigma/src/Enigma/Serialization/Json/JsonNumber.cs
@@ -5,9 +5,9 @@ namespace Enigma.Serialization.Json
 {
     public struct JsonNumber : IJsonNode
     {
-        public double? Value { get; }
+        public decimal? Value { get; }
 
-        public JsonNumber(double? value)
+        public JsonNumber(decimal? value)
         {
             Value = value;
         }
